requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
tensorflow
